cluster: rename block.py to block_id.py
